+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New config parameter 'useOidcClientAnnotation' to generate Quarkus OIDC client annotation for all resource interfaces.
 - New config parameter 'resourceProvidersOverride' to specify common set of Microprofile client providers for all resource interfaces.
 - New config parameter 'resourceConfigKeyOverride' to specify the same Microprofile client config-key for all resource interfaces.
+- Supporting 'style' og 'exploded' for Parameter annotations in resource interfaces.
 
 ### Changed
 - Bump dependency io.swagger.parser.v3:swagger-parser to v2.1.34